tools/py_stats/requirements.txt
absl_py==2.0.0
jupyterlab==4.0.9
matplotlib==3.8.2
numpy==1.26.2
pandas==2.1.3
Pillow==10.1.0
protobuf==4.25.1
pytest==7.4.3
